[TASK] #84610 | Implement feature against new submodule api
