--- Resources/Templates/Info.html
@@ -1,102 +1,0 @@
-{namespace ae=Psychomieze\AdminpanelExtended\ViewHelpers}
-<div class="container--tabs">
-	<section class="row">
-		<ul class="nav nav-tabs" data-typo3-role="typo3-adminPanel-tabs">
-			<li class="active" data-typo3-tab-target="general">General</li>
-			<li class="" data-typo3-tab-target="systemEnvironment">System Environment</li>
-			<li class="" data-typo3-tab-target="get">$_GET</li>
-			<li class="" data-typo3-tab-target="post">$_POST</li>
-			<li class="" data-typo3-tab-target="cookie">$_COOKIE</li>
-			<li class="" data-typo3-tab-target="server">$_SERVER</li>
-		</ul>
-		<div class="tab-content">
-			<div class="tab-pane active" data-typo3-role="typo3-adminPanel-tab" data-typo3-tab-id="general">
-				<div class="typo3-adminPanel-table-overflow">
-					<table class="typo3-adminPanel-table">
-						<f:if condition="{info.noCache}">
-							<f:render partial="Modules/Info/ImagesOnPage" arguments="{_all}" debug="false"/>
-							<f:render partial="Modules/Info/DocumentSize" arguments="{_all}" debug="false"/>
-							<tr>
-								<td>&nbsp;</td>
-								<td>&nbsp;</td>
-							</tr>
-						</f:if>
-						<f:render partial="Modules/Info/Page" arguments="{_all}" debug="false"/>
-						<f:render partial="Modules/Info/FeUserGroupList" arguments="{_all}" debug="false"/>
-						<f:render partial="Modules/Info/PageCache" arguments="{_all}" debug="false"/>
-						<f:render partial="Modules/Info/UserIntObjects" arguments="{_all}" debug="false"/>
-						<f:render partial="Modules/Info/FrontendUser" arguments="{_all}" debug="false"/>
-						<f:render partial="Modules/Info/TotalParsetime" arguments="{_all}" debug="false"/>
-					</table>
-				</div>
-			</div>
-			<div class="tab-pane" data-typo3-role="typo3-adminPanel-tab" data-typo3-tab-id="systemEnvironment">
-				<div class="typo3-adminPanel-table-overflow">
-					<table class="typo3-adminPanel-table">
-						<f:for each="{info.systemEnvironmentInformation}" as="info">
-							<tr>
-								<td>
-									<f:translate key="{info.title}" default="{info.title}"/>
-									<f:if condition="{info.titleAddition}"> ({info.titleAddition -> f:format.htmlspecialchars()})</f:if>
-								</td>
-								<td class="{f:if(condition: info.status, then: 'text-{info.status}')}">{info.value}</td>
-							</tr>
-						</f:for>
-					</table>
-
-					<hr>
-
-					<ae:phpInfo tableClass="typo3-adminPanel-table"/>
-				</div>
-			</div>
-			<div class="tab-pane" data-typo3-role="typo3-adminPanel-tab" data-typo3-tab-id="get">
-				<div class="typo3-adminPanel-table-overflow">
-					<table class="typo3-adminPanel-table">
-						<f:for each="{get}" as="val" key="key">
-							<tr>
-								<td>{key}</td>
-								<td>{val}</td>
-							</tr>
-						</f:for>
-					</table>
-				</div>
-			</div>
-			<div class="tab-pane" data-typo3-role="typo3-adminPanel-tab" data-typo3-tab-id="post">
-				<div class="typo3-adminPanel-table-overflow">
-					<table class="typo3-adminPanel-table">
-						<f:for each="{post}" as="val" key="key">
-							<tr>
-								<td>{key}</td>
-								<td>{val}</td>
-							</tr>
-						</f:for>
-					</table>
-				</div>
-			</div>
-			<div class="tab-pane" data-typo3-role="typo3-adminPanel-tab" data-typo3-tab-id="cookie">
-				<div class="typo3-adminPanel-table-overflow">
-					<table class="typo3-adminPanel-table">
-						<f:for each="{cookie}" as="val" key="key">
-							<tr>
-								<td>{key}</td>
-								<td>{val}</td>
-							</tr>
-						</f:for>
-					</table>
-				</div>
-			</div>
-			<div class="tab-pane" data-typo3-role="typo3-adminPanel-tab" data-typo3-tab-id="server">
-				<div class="typo3-adminPanel-table-overflow">
-					<table class="typo3-adminPanel-table">
-						<f:for each="{server}" as="val" key="key">
-							<tr>
-								<td>{key}</td>
-								<td>{val}</td>
-							</tr>
-						</f:for>
-					</table>
-				</div>
-			</div>
-		</div>
-	</section>
-</div>
