+++ Resources/Templates/Info/Environment.html
@@ -1,0 +1,11 @@
+<table class="typo3-adminPanel-table">
+	<f:for each="{systemEnvironmentInformation}" as="info">
+		<tr>
+			<td>
+				<f:translate key="{info.title}" default="{info.title}"/>
+				<f:if condition="{info.titleAddition}"> ({info.titleAddition -> f:format.htmlspecialchars()})</f:if>
+			</td>
+			<td class="{f:if(condition: info.status, then: 'text-{info.status}')}">{info.value}</td>
+		</tr>
+	</f:for>
+</table>
